+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ This document contains the FSL-MRS release history in reverse chronological orde
 - Add NIfTI-MRS reshape command
 - Add basis_tools remove_peak option to run HLSVD, typical usage for removing TMS peak.
 - Added an add_water_peak method to MRS class.
+- Updated fit_FSLModel defaults to match fsl_mrs command line defaults.
 
 1.1.3 (Tuesday 29th June 2021)
 ------------------------------